split the util.ts into two files to resolve circular imports
