+++ tfjs-core/src/environment.ts
@@ -16,7 +16,7 @@
  */
 
 import {Platform} from './platforms/platform';
-import {isPromise} from './util';
+import {isPromise} from './util_base';
 
 // Expects flags from URL in the format ?tfjsflags=FLAG1:1,FLAG2:true.
 const TENSORFLOWJS_FLAGS_PREFIX = 'tfjsflags';
